# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/synthesis": {
      "post": {
        "description": "Behind feature flag synthesis.enabled. Uses Bedrock when synthesis.provider.name=bedrock and the SDK adapter is wired; otherwise falls back to the deterministic provider. Refusals (flag off, no findings, low confidence) return 422.",
        "operationId": "SynthesisController_synthesize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Synthesize structured lookup findings into a brief plain-English narrative",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/admin/scim/tokens": {
      "get": {
        "operationId": "ScimTokenController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List SCIM bearer tokens for the calling tenant (no plaintext).",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "ScimTokenController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new SCIM bearer token. The plaintext is returned ONCE; store it now.",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/scim/tokens/{id}": {
      "delete": {
        "operationId": "ScimTokenController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Revoke a SCIM bearer token.",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Billing Rules Platform API",
    "description": "Pre-flight a claim against payer × state × code rules.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id"
      }
    },
    "schemas": {
...
      "SynthesizeDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTokenDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}